# Using ujson (if installed) in LazyJSON to loading json history 15% faster