Closes #4000

Test:
```python
pip uninstall ujson
XONSH_DEBUG=1 xonsh
# [history.json] Enumerated 2,852 history files for 0.7310s
# [history.json] Enumerated 2,852 history files for 0.7258s
# [history.json] Enumerated 2,852 history files for 0.6990s
# [history.json] Enumerated 2,852 history files for 0.7035s

pip install ujson
XONSH_DEBUG=1 xonsh
# [history.ujson] Enumerated 2,852 history files for 0.5832s
# [history.ujson] Enumerated 2,852 history files for 0.6084s
# [history.ujson] Enumerated 2,852 history files for 0.6103s
# [history.ujson] Enumerated 2,852 history files for 0.5903s
```